Revamp menus, split spec
Adopt a single menu which presents the different documents that are available.
Different versions of the same spec show in the menu.
The navigation menu expands the section being viewed.
The version selector is removed.

The introduction and how-to material are separated from the spec.

This introduces the build, build-env, and source tracks as separate specs that can be at different stage of maturity.

This also introduces v1.0.1 as a reformatted version of 1.0 which only contains core specification material and a pointer to the build track 1.0 spec .

Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>